requirements.txt
python-dotenv==0.20.0
python-telegram-bot==11.1.0
redis==3.2.1
